requirements.txt
supabase
google-generativeai
pymupdf
python-dotenv
flask
